A more proper fix for the bot/final client slot bug.

+++ source/game/p_weapon.c
@@ -1054,7 +1054,7 @@ void Blaster_Fire (edict_t *ent, vec3_t g_offset, int damage, qboolean hyper, in
 	}
 	else {
 		if(ent->client->buttons & BUTTON_ATTACK2) { //alt fire
-			fire_blaster_beam (ent, start, forward, damage/1.5, 0, false);
+			fire_blaster_beam (ent, start, forward, (int)((float)damage/1.5), 0, false);
 			gi.sound(ent, CHAN_AUTO, gi.soundindex("vehicles/shootlaser.wav"), 1, ATTN_NORM, 0);
 		}
 		else
